requirements.txt
fastapi==0.95.2
fastapi-elasticsearch==0.8.2
uvicorn==0.23.2
python-dotenv==0.20.0
python-multipart==0.0.6
requests==2.31.0